+++ client/src/components/desktop/Taskbar.tsx
@@ -1,47 +1,66 @@
-import { Button } from '@/components/ui/button'
-import { useState } from 'react'
-import { useDesktopStore } from '@/store/desktop'
-import { StartMenu } from './StartMenu'
-import { ContextMenu } from './ContextMenu'
-import { nanoid } from 'nanoid'
-import { AppWindow, Layout } from 'lucide-react'
-import { appIcons, getAppIcon } from '@/lib/appIcons'
+import { useEffect, useState } from 'react';
+import { Button } from '@/components/ui/button';
+import {
+  DropdownMenu,
+  DropdownMenuContent,
+  DropdownMenuItem,
+  DropdownMenuTrigger,
+} from '@/components/ui/dropdown-menu';
+import { Power, Maximize2 } from 'lucide-react';
+import { useDesktopStore } from '@/store/desktop';
+import { StartMenu } from './StartMenu';
+import { ContextMenu } from './ContextMenu';
+import { nanoid } from 'nanoid';
+import { AppWindow, Layout } from 'lucide-react';
+import { appIcons, getAppIcon } from '@/lib/appIcons';
 
 export function Taskbar() {
-  const { windows, activeWindowId, pinnedApps, setActiveWindow, toggleMinimize, addWindow, taskbarMode } = useDesktopStore()
-  const [showStartMenu, setShowStartMenu] = useState(false)
+  const { windows, activeWindowId, pinnedApps, setActiveWindow, toggleMinimize, addWindow, taskbarMode } = useDesktopStore();
+  const [showStartMenu, setShowStartMenu] = useState(false);
   const [contextMenu, setContextMenu] = useState<{
     x: number;
     y: number;
     appId: string;
     appComponent: string;
     appTitle: string;
-  } | null>(null)
+  } | null>(null);
+  const [dateTime, setDateTime] = useState('');
 
-  const handleRightClick = (
-    e: React.MouseEvent,
-    appId: string,
-    appComponent: string,
-    appTitle: string
-  ) => {
-    e.preventDefault()
+  useEffect(() => {
+    const updateTime = () => {
+      const now = new Date();
+      setDateTime(now.toLocaleString('en-US', {
+        weekday: 'short',
+        month: 'short',
+        day: 'numeric',
+        hour: 'numeric',
+        minute: '2-digit',
+        hour12: true
+      }));
+    };
+    updateTime();
+    const interval = setInterval(updateTime, 1000);
+    return () => clearInterval(interval);
+  }, []);
+
+  const handleRightClick = (e: React.MouseEvent, appId: string, appComponent: string, appTitle: string) => {
+    e.preventDefault();
     setContextMenu({
       x: e.clientX,
       y: e.clientY,
       appId,
       appComponent,
       appTitle,
-    })
-  }
+    });
+  };
 
   const handlePinnedAppClick = (app: { component: string; title: string }) => {
-    // Check if the app is already open
-    const existingWindow = windows.find((w) => w.component === app.component)
+    const existingWindow = windows.find((w) => w.component === app.component);
     if (existingWindow) {
       if (activeWindowId === existingWindow.id) {
-        toggleMinimize(existingWindow.id)
+        toggleMinimize(existingWindow.id);
       } else {
-        setActiveWindow(existingWindow.id)
+        setActiveWindow(existingWindow.id);
       }
     } else {
       addWindow({
@@ -56,13 +75,42 @@ export function Taskbar() {
         },
         isMinimized: false,
         isMaximized: false,
-      })
+      });
     }
-  }
+  };
+
+  const handleRestart = () => {
+    window.location.reload();
+  };
 
   return (
     <>
       <div className="fixed bottom-0 left-0 right-0 h-12 bg-background/80 backdrop-blur-md border-t flex items-center px-2 z-[9999]">
+        <DropdownMenu>
+          <DropdownMenuTrigger asChild>
+            <Button 
+              variant="ghost" 
+              size="icon" 
+              className="h-6 w-6 mr-2"
+            >
+              <Power className="h-4 w-4" />
+            </Button>
+          </DropdownMenuTrigger>
+          <DropdownMenuContent className="menu-transition">
+            <DropdownMenuItem onClick={handleRestart}>
+              Restart
+            </DropdownMenuItem>
+            <DropdownMenuItem onClick={() => {
+              document.body.classList.add('fade-out');
+              setTimeout(() => {
+                window.close();
+              }, 300);
+            }}>
+              Shut Down
+            </DropdownMenuItem>
+          </DropdownMenuContent>
+        </DropdownMenu>
+
         {taskbarMode !== 'windows11' && (
           <Button
             variant={showStartMenu ? "secondary" : "ghost"}
@@ -73,7 +121,7 @@ export function Taskbar() {
             <Layout className="h-5 w-5" />
           </Button>
         )}
-        
+
         <div className={`flex-1 flex items-center space-x-1 taskbar-content ${
           taskbarMode === 'chrome' || taskbarMode === 'windows11' ? 'justify-center' : ''
         }`}>
@@ -87,14 +135,13 @@ export function Taskbar() {
               <Layout className="h-5 w-5" />
             </Button>
           )}
-          {/* Pinned Apps */}
-            {pinnedApps.map((app) => {
-              const Icon = getAppIcon(app.component)
-              if (!Icon || !app.component || !app.title) {
-                return null
-              }
-              return (
-                <Button
+          {pinnedApps.map((app) => {
+            const Icon = getAppIcon(app.component);
+            if (!Icon || !app.component || !app.title) {
+              return null;
+            }
+            return (
+              <Button
                 key={app.component}
                 variant="ghost"
                 size="icon"
@@ -103,15 +150,14 @@ export function Taskbar() {
                 onContextMenu={(e) =>
                   handleRightClick(e, '', app.component, app.title)
                 }
-                >
+              >
                 <Icon className="h-4 w-4" />
-                </Button>
-              )
-            })}
+              </Button>
+            );
+          })}
           <div className="w-px h-6 bg-border mx-2" />
-          {/* Open Windows */}
           {windows.map((window) => {
-            const Icon = appIcons[window.component] || AppWindow
+            const Icon = appIcons[window.component] || AppWindow;
             return (
               <Button
                 key={window.id}
@@ -120,9 +166,9 @@ export function Taskbar() {
                 className="h-8 px-2 text-sm"
                 onClick={() => {
                   if (activeWindowId === window.id) {
-                    toggleMinimize(window.id)
+                    toggleMinimize(window.id);
                   } else {
-                    setActiveWindow(window.id)
+                    setActiveWindow(window.id);
                   }
                 }}
                 onContextMenu={(e) =>
@@ -132,11 +178,15 @@ export function Taskbar() {
                 <Icon className="h-4 w-4 mr-2" />
                 {window.title}
               </Button>
-            )
+            );
           })}
         </div>
+
+        <div className="absolute right-2 text-sm pointer-events-none select-none">
+          {dateTime}
+        </div>
       </div>
-      
+
       {showStartMenu && (
         <StartMenu onClose={() => setShowStartMenu(false)} />
       )}
@@ -148,5 +198,5 @@ export function Taskbar() {
         />
       )}
     </>
-  )
+  );
 }
